* refactoring * samples for creating tests from json object

+++ app/src/main/java/com/example/physmin/Pickable.kt
@@ -6,7 +6,7 @@ interface Pickable {
 
     var picked: Boolean
     var par: GroupPickable?
-    var answer: Short
+    var answer: Int
 
     fun setParent(_parent: GroupPickable)
